+++ src/extension.ts
@@ -104,12 +104,12 @@ export function activate(context: ExtensionContext) {
   /* load configuration and reload every time it's changed */
   logger.info(`Loading configuration...`);
   let config: Options = getOptions();
-  void Driver.setClient(config.useNativeDriver, context, poolingOptions(config));
+  Driver.clientReady = Driver.setClient(config.useNativeDriver, context, poolingOptions(config));
   context.subscriptions.push(
     workspace.onDidChangeConfiguration(() => {
       logger.debug("Configuration changed. Reloading configuration...");
       config = getOptions();
-      void Driver.setClient(config.useNativeDriver, context, poolingOptions(config));
+      Driver.clientReady = Driver.setClient(config.useNativeDriver, context, poolingOptions(config));
       commands.executeCommand("firebird.explorer.refresh");
     })
   );